Enable log-file debug
When booting it's good to print the log debug only if debug is enabled. However, it will be interesting to store this debug log in a file everytime as it consumes very low space. This is interesting to check the boot process without debug enabled, bacause in some cases the debug option fails to complete the boot. This patch changes the default behaviour to always write in the log file.