ssh.service: set correct file access
When the SSH key files comes from rootcopy dir the file access are incorrect. So it will be desirable to set the correct file access level (based on defaults) before starting the SSH service. Without this patch it's impossible to add some user custom keys to the rootcopy from a readonly storage (for example an ISO image). This fixes the problem.