Add user-preinit support (#181)
In some environments it will be interesting that the user can execute some script at the LiveKit level.

This patch enables the option to execute the script "/rootcopy/run/preinit.sh" just after ending the LiveKit setup process and before calling to the change_root. With this strategy the user can't break something releveant in the LiveKit but has the option to include some useful stuff before starting the regular boot with access to the entire filesystem. This could be used to test some develop to the LiveKit or to adapt the environment to custom requirements.